@@ -1,8 +1,10 @@
 name: example
 description: Example Project for flutter_chips_input
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.22.0"
 
 dependencies:
@@ -15,7 +17,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.2
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
