@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_chips_input/flutter_chips_input.dart';
 
 void main() {
-  final allContacts = const [
+  const allContacts = [
     'John Doe',
     'Jane Doe',
     'John Smith',
@@ -23,7 +23,7 @@ void main() {
                     .toList()
                 : const [],
             onChanged: (contacts) {
-              print(contacts);
+              debugPrint(contacts.toString());
             },
             chipBuilder: (context, state, contact) {
               return InputChip(
